requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-heroku==0.3.1
djangorestframework==3.9.1
psycopg2==2.8.3
pytz==2019.2
whitenoise==4.1.3